requirements.txt
rshell
esptool
micropython-esp32-stubs>=1.22.*
